+++ rs/rosetta-api/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for `NEURON_INFO` operation.
 - Support for `REMOVE_HOTKEY` operation.
+- Support for `FOLLOW` operation.
 
 ## [1.4.0] - 2022-03-14
 ### Added